requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
Deprecated==1.2.14
frozenlist==1.4.0
future==0.18.3
idna==3.4
line-bot-sdk==3.5.0
multidict==6.0.4
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.2
